refactor: move api and tests files to components folder
affects: @crave/farmblocks-link
