+++ packages/link/src/components/Link.test.js
@@ -2,7 +2,7 @@ import React from "react";
 import Enzyme, { shallow } from "enzyme";
 import Adapter from "enzyme-adapter-react-16";
 
-import Link from ".";
+import Link from "./Link";
 
 Enzyme.configure({ adapter: new Adapter() });
 describe("Link", function() {
