+++ packages/link/src/index.js
@@ -1,3 +1,3 @@
-export { default } from "./Link";
+export { default } from "./components/Link";
 export { fontSizes } from "@crave/farmblocks-theme";
 export { default as linkTypes } from "./constants/linkTypes";
